chore: ActualTableを切り出す
